@@ -29,24 +29,27 @@ namespace Personal.BlobStorage.Application
 
         private readonly IProductFileRepository _productFileRepository;
 
-        private readonly IFeatureManager _featureManager;
-        private readonly IConfigurationRefresher _refresher;
+        //private readonly IFeatureManager _featureManager;
+        //private readonly IConfigurationRefresher _refresher;
+        //private readonly IOptions<NameBasedUUIDOptions> _options;
 
         public BlobStorageContainer(ILogger<BlobStorageContainer> logger, IBlobClientUtilityService blobClientUtilityService, 
-            ILibraryUtilityService libraryUtilityService, IProductFileRepository productFileRepository, IFeatureManager featureManager, IConfigurationRefresher refresher)
+            ILibraryUtilityService libraryUtilityService, IProductFileRepository productFileRepository/*, IFeatureManager featureManager,*/
+            /*IConfigurationRefresher refresher, IOptions<NameBasedUUIDOptions> options*/)
         {
             _logger = logger;
             _blobClientUtilityService = blobClientUtilityService;
             _libraryUtilityService = libraryUtilityService;
             _productFileRepository = productFileRepository; 
-            _featureManager = featureManager;
-            _refresher = refresher;
+            //_featureManager = featureManager;
+            //_refresher = refresher;
+            //_options = options;
         }
 
 
         [HttpGet(Name = "GetWeatherForecast")]
         public async Task<IEnumerable<WeatherForecast>> Get()
-        {   
+        {
             // var sum = (int a, int b) => a + b;
 
             // Console.WriteLine(sum(1, 4));
@@ -86,21 +89,61 @@ namespace Personal.BlobStorage.Application
             //    Console.WriteLine(string.Join(",", item));
             //} 
 
+            
+
+            int[] numbers = { 5, 10, 8, 3, 6, 12 };
+
+            //Query syntax:
+            IEnumerable<int> numQuery1 =
+                from num in numbers
+                where num % 2 == 0
+                orderby num
+                select num;
+
+            //Method syntax:
+            IEnumerable<int> numQuery2 = numbers.Where(num => num % 2 == 0).OrderByDescending(n => n);
+
+            foreach (int i in numQuery1)
+            {
+                Console.Write(i + " ");
+            }
+
+            string str = "srikanth";
+            string str2 = "kan";
+            Console.WriteLine(str.IndexOf(str2));
+            Console.WriteLine(str.GetHashCode());
+            Console.WriteLine(System.Environment.NewLine);
+            foreach (int i in numQuery2)
+            {
+                Console.Write(i + " ");
+            }
+
+
             var coordinates = new Coords(1, 2);
             Console.WriteLine(coordinates);
-            await _refresher.TryRefreshAsync(new CancellationToken());
-            if (!await _featureManager.IsEnabledAsync("rise-assessments-service.rise-ftp-processor"))
-            {
-                //await _productFileRepository.UpsertItemAsync();
-                return Enumerable.Range(1, 5).Select(index => new WeatherForecast
-                {
-                    Date = DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
+            //await _refresher.TryRefreshAsync(new CancellationToken());
+            //if (!await _featureManager.IsEnabledAsync("rise-assessments-service.rise-ftp-processor"))
+            //{
+            //    //await _productFileRepository.UpsertItemAsync();
+            //    return Enumerable.Range(1, 5).Select(index => new WeatherForecast
+            //    {
+            //        Date = DateOnly.FromDateTime(DateTime.Now.AddDays(index)),
 
-                    TemperatureC = Random.Shared.Next(-20, 55),
-                    Summary = Summaries[Random.Shared.Next(Summaries.Length)]
-                })
-                .ToArray();
+            //        TemperatureC = Random.Shared.Next(-20, 55),
+            //        Summary = Summaries[Random.Shared.Next(Summaries.Length)]
+            //    })
+            //    .ToArray();
+            //}
+
+
+            List<string> names = ["srikanth", "Kranthi", "srikanth", "Chandhu", "srikanth"];
+            names = names.Distinct().ToList();  
+
+            foreach (var item in names.Distinct())
+            {
+                Console.WriteLine(item);
             }
+
             return null!;
 
             //var a = new []{1,2,3,4,5};
