@@ -3,15 +3,14 @@ namespace Personal.BlobStorage.Domain
 {
     public class BlobFileInfo
     {
-        public BlobFileInfo(string blobFileName, Dictionary<string, string>? metadata = null)
+        public BlobFileInfo(string blobFileName, Dictionary<string, string>? metaData = null)
         {
             Id = blobFileName ?? throw new ArgumentNullException(nameof(blobFileName));
             BlobFileName = blobFileName ?? throw new ArgumentNullException(nameof(blobFileName));
-            Metadata = metadata;
+            MetaData = metaData;
         }
-        [JsonPropertyName("id")]
         public string Id { get; set; }
         public string BlobFileName { get; set; }
-        public Dictionary<string, string>? Metadata { get; set; }
+        public Dictionary<string, string>? MetaData { get; set; }
     }
 }
